requirements.txt
asgiref==3.8.1
boto3==1.34.159
botocore==1.34.159
Django==5.1
django-environ==0.11.2
djangorestframework==3.15.2
jmespath==1.0.1
psycopg2-binary==2.9.9
pynamodb==6.0.1
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.2
